Fix lifetime issues for return value from pacrunner_proxy_lookup()
It is now a string which must be free'd with g_free().

A return of NULL will always be interpreted as "DIRECT", so no need to
return g_strdup("DIRECT");

This means that __pacrunner_mozjs_execute() is no longer returning a pointer
to a character array which may disappear at any time if GC happens.