curl: Handle recurring timeouts from curl
libcurl often fails to call our timer_callback function to set up the
next timer when it should. The can mean that a download hangs, for
example when it's waiting for getaddrinfo() to complete in a thread it
spawned for that purpose.

Don't ever allow the timer to be unset while there are downloads in
progress; use curl_multi_timeout() to check for the next timeout, and if
curl returns -1 to indicate that there's no timer pending, then set it
anyway for half a second.